Dixon - in libs/NcxxUtils, for 32-bit compat, in setDefaultFillvalue() for NC_LONG using missingInt

+++ codebase/libs/NcxxUtils/src/Ncxx/NcxxVar.cc
@@ -2445,7 +2445,7 @@ void NcxxVar::setDefaultFillvalue()
     return;
   }
   if (vtype == NC_LONG) {
-    addAttr("_fillValue", (long) Ncxx::missingInt);
+    addAttr("_fillValue", /* (long) */ Ncxx::missingInt);
     return;
   }
   if (vtype == NC_SHORT) {
